backend/requirements.txt
fastapi[standard]
pydantic-settings
tortoise-orm[asyncmy]
tortoise-orm[accel]
asyncmy
aerich[toml]
celery
chardet
arrow
apscheduler<4
loguru
redis
pyjwt
arrow
rich
aiohttp
orjson<4
pydantic<3
arrow
